adding more tests based on review comments
