+++ src/test/groovy/io/openliberty/tools/gradle/BaseDevTest.groovy
@@ -258,6 +258,26 @@ class BaseDevTest extends AbstractIntegrationTest {
         Files.write(path, content.getBytes(charset));
     }
 
+    /**
+    * Count number of lines that contain the given string
+    */
+    protected static int countOccurrences(String str, File file) throws FileNotFoundException, IOException {
+        int occurrences = 0;
+        BufferedReader br = new BufferedReader(new FileReader(file));
+        String line = br.readLine();
+        try {
+            while (line != null) {
+                if (line.contains(str)) {
+                    occurrences++;
+                }
+                line = br.readLine();
+            }
+        } finally {
+            br.close();
+        }
+        return occurrences;
+    }
+
     protected static void cleanUpAfterClass(boolean isDevMode) throws Exception {
         stopProcess(isDevMode, errFile);
         if (buildDir != null && buildDir.exists()) {
